interface/isldlname.py: create destination directory
An upcoming commit will store (the machine independent part of)
the Python interface in the git repository.
This means the interface will be available even if the 'interface'
subdirectory is not being built.  This directory may therefore
not exist when the machine dependent part of the interface
is being constructed, so create it if needed.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>