isl_basic_map_remove_shifted_constraints: align with context first
The function isl_basic_map_align_divs currently does not support
existentially quantified variables in the basic map that prescribes
the order of the local variables.
This will be allowed in a future commit, but this will result
in additional existentially quantified variables being introduced
in the other basic map.

Since the context is known not to involve any existentially
quantified variables, align with it first to avoid
additional existentially quantified variables being introduced twice.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>