check for proper set spaces where needed
A space can have 0, 1 or 2 tuples.
The first two can be used as the space for an isl_set,
while the last can be used as the space of an isl_map.
There is no separate type for a relation with zero tuples.
The function isl_space_check_is_set therefore succeeds
on spaces with either 0 or 1 tuple.

In some cases, however, a (single) tuple is required.
Use isl_space_check_is_proper_set to check
that the space has exactly one tuple where needed.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>