fix printing of empty piecewise expressions
An empty piecewise expression would simply get printed as "{  }",
but this makes it impossible to recover the space of
the piecewise expression.
Print a dummy base element with all zeros for the dimensions,
followed by the constraints of an empty domain instead.
This is similar to how an empty isl_pw_aff inside an isl_multi_pw_aff
is printed as "(0 : false)".

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>