only copy generated interfaces to source tree when needed
Since isl-0.26-13-g55d9c2394e (generate interfaces in source tree,
Sat Mar 18 17:45:40 2023 +0100), the generated interfaces
are stored in the source tree so that they can also be stored
in the git repository.

In general, it is not a good idea to modify the source tree
when building out of the source tree.
In particular, the source tree may be read-only.
Therefore, only copy the interfaces to the source tree
when it is actually needed for updating the git repository,
i.e., when the generated interfaces are different from
the ones currently in the source tree.

This means the timestamps of the files in the source tree
will not be updated and so the rule for making these files
will trigger every time the files are needed.
The files are therefore first generated in the build tree and
then copied to the source tree when needed.
This compare-and-copy should be fairly quick and is therefore
made completely silent by default.

Note that if it is only the files themselves that are read-only
then it may still be possible to use "mv -f" or to touch the files,
but the current solution is more general.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>