move Wrap helper class to installed isl-interface/clang_wrap.h
This allows the helper class to be reused by pet.

The reusable code is extracted as a pure header
rather than a library to reduce the risk of complications.
In particular, extract_interface is run on the build system,
while pet is run on the host system, so it would not be possible
to share a single library in general.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>