clang_wrap.h: do not return reference to local object in getFile
The getFileRef method returns an llvm::Expected<clang::FileEntryRef>, and
operator* on an llvm::Expected returns a reference to the embedded object,
which is destroyed when getFile returns.  This problem is usually masked by
compiler optimizations.

Return the llvm::Expected<clang::FileEntryRef> object itself.

Note that the caller does not need to be adjusted because
an llvm::ErrorOr<clang::FileEntryRef> behaves similarly
to an llvm::Expected in the sense that they both evaluate
to false in the case of an error and they can both
be dereferenced to obtain the embedded object.

Reported-by: Leo Yin <yinle2025@gmail.com>
Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>