interface: adjust getFileRef method detection for compatibility with gcc 4.9
Commit isl-0.26-152-g146c2f6aa8 (interface: update to removal of
SourceManager::createFileID(FileEntry*), Sat Jul 6 18:34:27 2024 +0200)
introduced a mechanism for detecting the presence of
the getFileRef method in FileManager that should be valid C++11,
but that apparently does not work in gcc 4.9.

Use a different mechanism that moreover checks for a specific overload
rather than just the presence of some getFileRef method.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>